# resources/openAPI/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Dynamic Mapper API",
    "description" : "REST API for managing Dynamic Mapper configurations including connectors, mappings,\nservice settings, and code templates. The Dynamic Mapper enables bi-directional\ndata transformation between external systems and Cumulocity IoT.\n\n**Key Features:**\n- Connector management (MQTT, HTTP, TCP, etc.)\n- Service configuration and feature flags\n- Code template management for custom processing\n- Multi-tenant support with role-based access control\n",
    "contact" : {
      "name" : "Cumulocity Dynamic Mapper Team",
      "email" : "support@cumulocity.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version" : "v1.0"
  },
...
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "basicAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/relation/client/{clientId}" : {
      "put" : {
        "tags" : [ "Client Relation Controller" ],
        "summary" : "Update client relation for client",
        "operationId" : "addOrUpdateClientRelations",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Client relation updated successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - insufficient permissions",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Device not found or outbound relation disabled",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Client Relation Controller" ],
        "summary" : "Remove all device mappings for a specific client",
        "operationId" : "removeAllRelationsForClient",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "All mappings for client removed successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - insufficient permissions",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Client not found or outbound mapping disabled",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/configuration/service" : {
      "get" : {
        "tags" : [ "Configuration Management" ],
        "summary" : "Get service configuration",
        "description" : "Retrieves the service configuration for the current tenant including feature flags, cache settings, and operational parameters.",
        "operationId" : "getServiceConfiguration",
        "responses" : {
          "200" : {
            "description" : "Service configuration retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceConfiguration"
                }
              }
            }
          },
          "404" : {
            "description" : "Service configuration not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      },
      "put" : {
        "tags" : [ "Configuration Management" ],
        "summary" : "Update service configuration",
        "description" : "Updates the service configuration for the current tenant.\n\n**Important:** Changing outbound mapping settings will affect notification subscriptions\nand may trigger connector reconnections.\n\n**Security:** Requires `ROLE_DYNAMIC_MAPPER_ADMIN`\n",
        "operationId" : "updateServiceConfiguration",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "additionalProperties" : {
                  "type" : "object"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Service configuration updated successfully"
          },
          "400" : {
            "description" : "Invalid service configuration"
          },
          "403" : {
            "description" : "Insufficient permissions"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
...
    "/relation/device/{deviceId}/client" : {
      "get" : {
        "tags" : [ "Client Relation Controller" ],
        "summary" : "Get client mapping for device",
        "operationId" : "getClientForDevice",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Client mapping retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Device not found, client mapping not found, or outbound mapping disabled",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/relation/clients" : {
      "get" : {
        "tags" : [ "Client Relation Controller" ],
        "summary" : "Get all clients",
        "operationId" : "getAllClients",
        "responses" : {
          "200" : {
            "description" : "All clients retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Outbound mapping disabled",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/relation/client" : {
      "get" : {
        "tags" : [ "Client Relation Controller" ],
        "summary" : "Get all client relations",
        "operationId" : "getAllClientRelations",
        "responses" : {
          "200" : {
            "description" : "All client relations retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Outbound mapping disabled",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Client Relation Controller" ],
        "summary" : "Clear all client relations",
        "operationId" : "clearAllClientRelations",
        "responses" : {
          "200" : {
            "description" : "All client relations cleared successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - insufficient permissions",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Outbound mapping disabled",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/relation/client/{clientId}/devices" : {
      "get" : {
        "tags" : [ "Client Relation Controller" ],
        "summary" : "Get all devices mapped to a specific client",
        "operationId" : "getDevicesForClient",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Devices for client retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Client not found or outbound mapping disabled",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/relation/device/{deviceId}" : {
      "delete" : {
        "tags" : [ "Client Relation Controller" ],
        "summary" : "Remove client mapping for specific device",
        "operationId" : "removeRelationForDevice",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Client mapping deleted successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - insufficient permissions",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Device not found, client mapping not found, or outbound mapping disabled",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
...
    "securitySchemes" : {
      "basicAuth" : {
        "type" : "http",
        "description" : "Basic Authentication using Cumulocity IoT credentials",
        "in" : "header",
        "scheme" : "basic"
      }
    }
  }
}